+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Delete functionality for characters and media content
+
 ## [v1.0.0] - 2025-01-12
 
 ### Added